+++ src/Mod/Fem/App/FemMeshPyImp.cpp
@@ -575,8 +575,8 @@ PyObject* FemMeshPy::getNodesByEdge(PyObject *args)
             return 0;
         }
         Py::List ret;
-        std::set<long> resultSet = getFemMeshPtr()->getNodesByEdge(fc);
-        for (std::set<long>::const_iterator it = resultSet.begin();it!=resultSet.end();++it)
+        std::set<int> resultSet = getFemMeshPtr()->getNodesByEdge(fc);
+        for (std::set<int>::const_iterator it = resultSet.begin();it!=resultSet.end();++it)
             ret.append(Py::Int(*it));
 
         return Py::new_reference_to(ret);
